db: move flushable and flushableEntry to their own file
