+++ db.go
@@ -40,63 +40,6 @@ var (
 	ErrReadOnly = errors.New("pebble: read-only")
 )
 
-type flushable interface {
-	newIter(o *IterOptions) internalIterator
-	newFlushIter(o *IterOptions, bytesFlushed *uint64) internalIterator
-	newRangeDelIter(o *IterOptions) internalIterator
-	inuseBytes() uint64
-	totalBytes() uint64
-	readyForFlush() bool
-}
-
-type flushableEntry struct {
-	flushable
-	// Channel which is closed when the flushable has been flushed.
-	flushed chan struct{}
-	// flushForced indicates whether a flush was forced on this memtable (either
-	// manual, or due to ingestion). Protected by DB.mu.
-	flushForced bool
-	// logNum corresponds to the WAL that contains the records present in the
-	// receiver.
-	logNum uint64
-	// logSize is the size in bytes of the associated WAL. Protected by DB.mu.
-	logSize uint64
-	// The current logSeqNum at the time the memtable was created. This is
-	// guaranteed to be less than or equal to any seqnum stored in the memtable.
-	logSeqNum uint64
-	// readerRefs tracks the read references on the flushable. The two sources of
-	// reader references are DB.mu.mem.queue and readState.memtables. The memory
-	// reserved by the flushable in the cache is released when the reader refs
-	// drop to zero. If the flushable is a memTable, when the reader refs drops
-	// to zero, the writer refs will already be zero because the memtable will
-	// have been flushed and that only occurs once the writer refs drops to zero.
-	readerRefs int32
-	// Closure to invoke to release memory accounting.
-	releaseMemAccounting func()
-}
-
-func (e *flushableEntry) readerRef() {
-	switch v := atomic.AddInt32(&e.readerRefs, 1); {
-	case v <= 1:
-		panic(fmt.Sprintf("pebble: inconsistent reference count: %d", v))
-	}
-}
-
-func (e *flushableEntry) readerUnref() {
-	switch v := atomic.AddInt32(&e.readerRefs, -1); {
-	case v < 0:
-		panic(fmt.Sprintf("pebble: inconsistent reference count: %d", v))
-	case v == 0:
-		if e.releaseMemAccounting == nil {
-			panic(fmt.Sprintf("pebble: memtable reservation already released"))
-		}
-		e.releaseMemAccounting()
-		e.releaseMemAccounting = nil
-	}
-}
-
-type flushableList []*flushableEntry
-
 // Reader is a readable key/value store.
 //
 // It is safe to call Get and NewIter from concurrent goroutines.
